third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_continue_objectstore3-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.continue() - object store - attempt to iterate to the previous record when the direction is set for the next record 
Harness: the test ran to completion.
